+++ app/src/main/java/com/techcourse/AppWebApplicationInitializer.java
@@ -1,5 +1,8 @@
 package com.techcourse;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
 import jakarta.servlet.ServletContext;
 import nextstep.mvc.DispatcherServlet;
 import nextstep.mvc.HandlerAdapterRegistry;
@@ -7,11 +10,6 @@ import nextstep.mvc.HandlerMappingRegistry;
 import nextstep.mvc.handler.adapter.AnnotationHandlerAdapter;
 import nextstep.mvc.handler.mapping.AnnotationHandlerMapping;
 import nextstep.web.WebApplicationInitializer;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-import com.techcourse.support.web.handler.ManualHandlerAdapter;
-import com.techcourse.support.web.handler.ManualHandlerMapping;
 
 public class AppWebApplicationInitializer implements WebApplicationInitializer {
 
@@ -20,12 +18,10 @@ public class AppWebApplicationInitializer implements WebApplicationInitializer {
     @Override
     public void onStartup(final ServletContext servletContext) {
         final var handlerMappingRegistry = HandlerMappingRegistry.builder()
-            .add(new ManualHandlerMapping())
             .add(new AnnotationHandlerMapping("com.techcourse"))
             .build();
 
         final var handlerAdapterRegistry = HandlerAdapterRegistry.builder()
-            .add(new ManualHandlerAdapter())
             .add(new AnnotationHandlerAdapter())
             .build();
