+++ app/src/main/java/com/techcourse/controller/LoginController.java
@@ -4,20 +4,23 @@ import com.techcourse.domain.User;
 import com.techcourse.repository.InMemoryUserRepository;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-import nextstep.mvc.controller.asis.Controller;
+import nextstep.web.annotation.Controller;
+import nextstep.web.annotation.RequestMapping;
+import nextstep.web.support.RequestMethod;
+
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-public class LoginController implements Controller {
+@Controller
+public class LoginController {
 
     private static final Logger log = LoggerFactory.getLogger(LoginController.class);
 
-    @Override
-    public String execute(final HttpServletRequest req, final HttpServletResponse res) throws Exception {
+    @RequestMapping(value = "/login", method = RequestMethod.POST)
+    public String login(final HttpServletRequest req, final HttpServletResponse res) throws Exception {
         if (UserSession.isLoggedIn(req.getSession())) {
             return "redirect:/index.jsp";
         }
-
         return InMemoryUserRepository.findByAccount(req.getParameter("account"))
                 .map(user -> {
                     log.info("User : {}", user);
@@ -31,8 +34,24 @@ public class LoginController implements Controller {
             final var session = request.getSession();
             session.setAttribute(UserSession.SESSION_KEY, user);
             return "redirect:/index.jsp";
-        } else {
-            return "redirect:/401.jsp";
         }
+        return "redirect:/401.jsp";
+    }
+
+    @RequestMapping(value = "/login", method = RequestMethod.GET)
+    public String loginView(final HttpServletRequest req, final HttpServletResponse res) throws Exception {
+        return UserSession.getUserFrom(req.getSession())
+            .map(user -> {
+                log.info("logged in {}", user.getAccount());
+                return "redirect:/index.jsp";
+            })
+            .orElse("/login.jsp");
+    }
+
+    @RequestMapping(value = "/logout", method = RequestMethod.GET)
+    public String execute(final HttpServletRequest req, final HttpServletResponse res) throws Exception {
+        final var session = req.getSession();
+        session.removeAttribute(UserSession.SESSION_KEY);
+        return "redirect:/";
     }
 }
