--- app/src/main/java/com/techcourse/support/web/handler/ManualHandlerAdapter.java
@@ -1,23 +1,0 @@
-package com.techcourse.support.web.handler;
-
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import nextstep.mvc.handler.adapter.HandlerAdapter;
-import nextstep.mvc.controller.asis.Controller;
-import nextstep.mvc.view.JspView;
-import nextstep.mvc.view.ModelAndView;
-
-public class ManualHandlerAdapter implements HandlerAdapter {
-
-	@Override
-	public boolean supports(Object handler) {
-		return handler instanceof Controller;
-	}
-
-	@Override
-	public ModelAndView handle(HttpServletRequest request, HttpServletResponse response, Object handler) throws Exception {
-		Controller controller = (Controller)handler;
-		String viewName = controller.execute(request, response);
-		return new ModelAndView(new JspView(viewName));
-	}
-}
