--- app/src/main/java/com/techcourse/support/web/handler/ManualHandlerMapping.java
@@ -1,39 +1,0 @@
-package com.techcourse.support.web.handler;
-
-import com.techcourse.controller.*;
-import jakarta.servlet.http.HttpServletRequest;
-import nextstep.mvc.handler.mapping.HandlerMapping;
-import nextstep.mvc.controller.asis.Controller;
-import nextstep.mvc.controller.asis.ForwardController;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-import java.util.HashMap;
-import java.util.Map;
-import java.util.Optional;
-
-public class ManualHandlerMapping implements HandlerMapping {
-
-    private static final Logger log = LoggerFactory.getLogger(ManualHandlerMapping.class);
-
-    private static final Map<String, Controller> controllers = new HashMap<>();
-
-    @Override
-    public void initialize() {
-        controllers.put("/", new ForwardController("/index.jsp"));
-        controllers.put("/login", new LoginController());
-        controllers.put("/login/view", new LoginViewController());
-        controllers.put("/logout", new LogoutController());
-
-        log.info("Initialized Handler Mapping!");
-        controllers.keySet()
-                .forEach(path -> log.info("Path : {}, Controller : {}", path, controllers.get(path).getClass()));
-    }
-
-    @Override
-    public Optional<Object> getHandler(HttpServletRequest request) {
-        final String requestURI = request.getRequestURI();
-        log.debug("Request Mapping Uri : {}", requestURI);
-        return Optional.ofNullable(controllers.get(requestURI));
-    }
-}
